Update documentation to reflect the steps required  to set up a node
This commit updates the documentation to remove the use of the
cleveldb requirement, as the foundation has advised me this has been
removed following version 1.13. I've also refreshed the docs for
both mainnet and testnet to indicate that these were written under
the assumption of v1.16.0 as there were multiple version references
in these pages. I've also described the necessity of using the `-t`
flag when running a testnet node, as this can lead to undefined
behaviour if not supplied when running the testnet chain.